Added gradual speed change which affects score and FOV
Also tidied up the gameplay scene a little bit and fixed an issue of the fps limit not being displayed properly at first on the settings button.

+++ src/Game.h
@@ -78,7 +78,7 @@ private:
 	bool mSceneChanged = false;
 
 	bool mLimitingFPS = false;
-	float mFPSLimit = 60;
+	int mFPSLimit = 60;
 
 	// To detect changes for shader uniforms
 	glm::mat4 mLastProjection = glm::mat4(1.f);
